requirements.txt
ddgs==9.5.5
duckduckgo_search==8.1.1
fastapi==0.116.1
langchain==0.3.27
langchain_core==0.3.76
langchain_openai==0.3.33
langgraph==0.6.7
matplotlib==3.10.6
mysql_connector_python==8.0.32
numpy==2.3.3
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0
